refactor: Factor out note scripts
