+++ miden-lib/src/notes/mod.rs
@@ -12,7 +12,9 @@ use miden_objects::{
     vm::Program,
     Felt, NoteError, Word,
 };
+use utils::build_swap_tag;
 
+pub mod scripts;
 pub mod utils;
 
 // STANDARDIZED SCRIPTS
@@ -36,10 +38,7 @@ pub fn create_p2id_note<R: FeltRng>(
     aux: Felt,
     rng: &mut R,
 ) -> Result<Note, NoteError> {
-    let bytes = include_bytes!(concat!(env!("OUT_DIR"), "/assets/note_scripts/P2ID.masb"));
-    let program =
-        Program::read_from_bytes(bytes).map_err(NoteError::NoteScriptDeserializationError)?;
-    let note_script = NoteScript::new(program);
+    let note_script = scripts::p2id();
 
     let inputs = NoteInputs::new(vec![target.into()])?;
     let tag = NoteTag::from_account_id(target, NoteExecutionMode::Local)?;
@@ -72,10 +71,7 @@ pub fn create_p2idr_note<R: FeltRng>(
     recall_height: u32,
     rng: &mut R,
 ) -> Result<Note, NoteError> {
-    let bytes = include_bytes!(concat!(env!("OUT_DIR"), "/assets/note_scripts/P2IDR.masb"));
-    let program =
-        Program::read_from_bytes(bytes).map_err(NoteError::NoteScriptDeserializationError)?;
-    let note_script = NoteScript::new(program);
+    let note_script = scripts::p2idr();
 
     let inputs = NoteInputs::new(vec![target.into(), recall_height.into()])?;
     let tag = NoteTag::from_account_id(target, NoteExecutionMode::Local)?;
@@ -104,10 +100,7 @@ pub fn create_swap_note<R: FeltRng>(
     aux: Felt,
     rng: &mut R,
 ) -> Result<(Note, NoteDetails), NoteError> {
-    let bytes = include_bytes!(concat!(env!("OUT_DIR"), "/assets/note_scripts/SWAP.masb"));
-    let program =
-        Program::read_from_bytes(bytes).map_err(NoteError::NoteScriptDeserializationError)?;
-    let note_script = NoteScript::new(program);
+    let note_script = scripts::swap();
 
     let payback_serial_num = rng.draw_word();
     let payback_recipient = utils::build_p2id_recipient(sender, payback_serial_num)?;
@@ -144,41 +137,4 @@ pub fn create_swap_note<R: FeltRng>(
     let payback_note = NoteDetails::new(payback_assets, payback_recipient);
 
     Ok((note, payback_note))
-}
-
-// HELPER FUNCTIONS
-// ================================================================================================
-
-/// Returns a note tag for a swap note with the specified parameters.
-///
-/// Use case ID for the returned tag is set to 0.
-///
-/// Tag payload is constructed by taking asset tags (8 bits of faucet ID) and concatenating them
-/// together as offered_asset_tag + requested_asset tag.
-///
-/// Network execution hint for the returned tag is set to `Local`.
-pub fn build_swap_tag(
-    note_type: NoteType,
-    offered_asset: &Asset,
-    requested_asset: &Asset,
-) -> Result<NoteTag, NoteError> {
-    const SWAP_USE_CASE_ID: u16 = 0;
-
-    // get bits 4..12 from faucet IDs of both assets, these bits will form the tag payload; the
-    // reason we skip the 4 most significant bits is that these encode metadata of underlying
-    // faucets and are likely to be the same for many different faucets.
-
-    let offered_asset_id: u64 = offered_asset.faucet_id().into();
-    let offered_asset_tag = (offered_asset_id >> 52) as u8;
-
-    let requested_asset_id: u64 = requested_asset.faucet_id().into();
-    let requested_asset_tag = (requested_asset_id >> 52) as u8;
-
-    let payload = ((offered_asset_tag as u16) << 8) | (requested_asset_tag as u16);
-
-    let execution = NoteExecutionMode::Local;
-    match note_type {
-        NoteType::Public => NoteTag::for_public_use_case(SWAP_USE_CASE_ID, payload, execution),
-        _ => NoteTag::for_local_use_case(SWAP_USE_CASE_ID, payload),
-    }
 }
